refactor(board-mixins): clean up controls mixin styles and a bit of cleanup
