--- packages/components/src/board-mixins/mixin-controls.st.css
@@ -1,50 +1,0 @@
-:vars {
-  transition: all 1s;
-}
-
-.root{
-  background: white;
-  border-radius: 5px;
-  padding: 3px;
-  position: fixed;
-  bottom: 30px;
-  right: 30px;
-  box-shadow: 2px 2px 10px #898989;
-  width: 40px;
-  height: 40px;
-  border-radius: 50%;
-  overflow: hidden;
-  transition: value(transition);
-  font: 14px arial;
-}
-.root::before{
-  content: "actions";
-  background: white;
-  top: 30%;
-  left: 0px;
-  height: 20px;
-  position: absolute;
-  transition: value(transition);
-}
-.root:hover::before{
-  content: "";
-  height: 0px;
-  opacity: 0;
-  transition: value(transition);
-}
-.root:hover{
-  width: auto;
-  height: auto;
-  border-radius: 0px;
-  transition: value(transition);
-}
-.root button{
-  background-color: cornflowerblue;
-  color: white;
-  border-radius: 5px;
-  padding: 5px;
-  cursor: pointer;
-}
-.root button:hover{
-  background-color: cadetblue;
-}
