+++ packages/components/src/board-mixins/mixin-controls/mixin-controls.st.css
@@ -1,0 +1,57 @@
+.root {
+    position: fixed;
+    z-index: 999;
+
+    right: 30px;
+    bottom: 30px;
+
+    width: 40px;
+    height: 40px;
+    overflow: hidden;
+
+    background: white;
+    padding: 3px;
+    box-shadow: 2px 2px 10px #898989;
+
+    font: 14px arial;
+}
+
+.root::before {
+    content: "actions";
+
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    height: 100%;
+    width: 100%;
+
+    position: absolute;
+    top: 0;
+    left: 0;
+
+    background: white;
+}
+
+.root:hover::before {
+    display: none;
+}
+
+.root:hover {
+    width: 50%;
+    height: auto;
+    display: flex;
+    flex-direction: column;
+    gap: 8px;
+}
+
+.root button {
+    background-color: cornflowerblue;
+    color: white;
+    border-radius: 5px;
+    padding: 5px;
+    cursor: pointer;
+}
+
+.root button:hover {
+    background-color: cadetblue;
+}
